# Avoid unnecessary ToArray()