Unnecessary ToArray() causes significant extra memory allocation.
Avoid full double enumeration by calling pixels.Take(2).Count().