@swharden 2024-04-06
I just created a benchmark project (#3190) and testing this PR is its first official use 🚀 

These tests evaluate rendering a data logger render with different numbers of points:

## Before
| Method           | Points | Mean       | Error | Gen0    | Allocated  |
|----------------- |------- |-----------:|------:|--------:|-----------:|
| DataLoggerRender | 100    |   1.212 ms |    NA | 30.4878 |  290.21 KB |
| DataLoggerRender | 1000   |   4.996 ms |    NA | 26.3158 |  298.49 KB |
| DataLoggerRender | 10000  |  34.621 ms |    NA |       - |  365.04 KB |
| DataLoggerRender | 100000 | 298.433 ms |    NA |       - | 1073.73 KB |

## After this PR
| Method           | Points | Mean       | Error | Gen0    | Allocated |
|----------------- |------- |-----------:|------:|--------:|----------:|
| DataLoggerRender | 100    |   1.244 ms |    NA | 31.4465 | 289.43 KB |
| DataLoggerRender | 1000   |   4.944 ms |    NA | 25.6410 | 290.67 KB |
| DataLoggerRender | 10000  |  35.664 ms |    NA |       - | 286.91 KB |
| DataLoggerRender | 100000 | 295.924 ms |    NA |       - | 292.48 KB |